fix delete character
Former-commit-id: 9be4bb617232f1723e6bbb00022133d72a736401

+++ proto/client/cmsg.go
@@ -180,7 +180,7 @@ type NewCharacter struct {
 }
 
 type DeleteCharacter struct {
-	CharacterIndex int16
+	CharacterIndex int32
 }
 
 type StartGame struct {
